Add widgets for form fields
These are based on the tags field, with local modifications (don't allow
free values, reduce interval, support action API like responses). Values
are validated and transformed to URIs for internal storage, and
transformed to titles on the frontend.

Also added two snippets for displaying the values on the dataset page